♻️ split agent runtime libraries by responsibility
