+++ libs/backend/agents/runtime/README.md
@@ -44,6 +44,10 @@ composes it. Cleanup implements the execution/runs physical store contract struc
 the backend-to-infrastructure dependency direction. Runtime packages do not import a model driver.
 Canonical run/event persistence remains in its owning backend domain.
 
+Each package barrel is a composition boundary: it exposes runnable factories and required policy
+types, while reconciliation seams, Kubernetes client ports, response decoders, and manifest
+protocol details remain owned inside their package.
+
 ## See also
 
 - Parent group: [agents](../README.md)
