+++ libs/backend/agents/runtime/cleanup/main/src/index.ts
@@ -1,2 +1,1 @@
 export { __CreateKubernetesRuntimeWorkloadCleanupStore } from "./runtime-workload-cleanup-store.js";
-export type { KubernetesRuntimeWorkloadCleanupBatchApi, KubernetesRuntimeWorkloadCleanupProjection, KubernetesRuntimeWorkloadCleanupResult, KubernetesRuntimeWorkloadCleanupStore, KubernetesRuntimeWorkloadCleanupStoreOptions } from "./runtime-workload-cleanup-store.types.js";
