+++ libs/backend/agents/runtime/controller/src/agent-controller.types.ts
@@ -1,4 +1,4 @@
-import type { ConfigurationOptions, V1Job, V1Pod, V1PodList, V1Secret } from "@kubernetes/client-node";
+import type { V1Job, V1Pod, V1Secret } from "@kubernetes/client-node";
 import type { Logger } from "@opencrane/observability";
 import type { AgentControllerRunAttemptAssignmentCommand, AgentControllerRunAttemptAssignmentResult, AgentControllerRunAttemptClaim, AgentControllerRunWorkloadRegistrationCommand, AgentControllerRunWorkloadRegistrationResult, AgentControllerRunWorkloadReleaseClaim } from "@opencrane/contracts";
 import type { AgentRuntimeJobProfile } from "@opencrane/backend/agents/runtime/k8s-launcher";
@@ -65,87 +65,20 @@ export interface AgentControllerOptions
 
 /** Result of one desired-state poll. */
 export type AgentControllerReconcileResult =
-	| { readonly outcome: "idle" }
+	| { readonly outcome: AgentControllerReconcileOutcomes.Idle }
 	| { readonly outcome: "assigned" | "idempotent"; readonly eventId: string; readonly runId: string; readonly attempt: number; readonly workloadUid: string };
+
+/** Controller-owned loop outcomes that determine whether another poll may be delayed. */
+export enum AgentControllerReconcileOutcomes
+{
+	/** No durable work was claimable. */
+	Idle = "idle",
+	/** Release succeeded but Kubernetes has not created the uniquely owned first Pod. */
+	PendingPod = "pending-pod",
+}
 
 /** Result of one workload-release poll. */
 export type AgentControllerRuntimeReleaseReconcileResult =
-	| { readonly outcome: "idle" }
-	| { readonly outcome: "pending-pod"; readonly eventId: string; readonly runId: string; readonly attempt: number; readonly workloadUid: string }
+	| { readonly outcome: AgentControllerReconcileOutcomes.Idle }
+	| { readonly outcome: AgentControllerReconcileOutcomes.PendingPod; readonly eventId: string; readonly runId: string; readonly attempt: number; readonly workloadUid: string }
 	| { readonly outcome: "registered" | "idempotent"; readonly eventId: string; readonly runId: string; readonly attempt: number; readonly workloadUid: string; readonly podUid: string };
-
-/** Fetch-compatible function injected into the HTTP adapter. */
-export type AgentControllerFetch = (input: string | URL | Request, init?: RequestInit) => Promise<Response>;
-
-/** Rotating projected-token reader injected into the HTTP adapter. */
-export type AgentControllerTokenReader = () => Promise<string>;
-
-/** Configuration for the projected-token-authenticated OpenCrane adapter. */
-export interface AgentControllerHttpAuthorityOptions
-{
-	/** Internal OpenCrane base URL with no path, query, or credentials. */
-	readonly openCraneInternalUrl: string;
-	/** Absolute path of the rotating projected controller token. */
-	readonly tokenPath: string;
-	/** Hard timeout for one HTTP exchange. */
-	readonly requestTimeoutMilliseconds: number;
-	/** Optional fetch seam used by focused tests. */
-	readonly fetch?: AgentControllerFetch;
-	/** Optional rotating-token seam used by focused tests. */
-	readonly readToken?: AgentControllerTokenReader;
-}
-
-/** Narrow Batch API surface used for exact Job creation, reads, and release. */
-export interface AgentControllerBatchApi
-{
-	/** Create one suspended namespaced Job. */
-	createNamespacedJob(request: { readonly namespace: string; readonly body: V1Job }, options?: ConfigurationOptions): Promise<V1Job>;
-	/** Read one deterministic Job after an AlreadyExists response. */
-	readNamespacedJob(request: { readonly namespace: string; readonly name: string }, options?: ConfigurationOptions): Promise<V1Job>;
-	/** Apply one conditional JSON Patch to the exact assigned Job. */
-	patchNamespacedJob(request: AgentControllerJobPatchRequest, options?: ConfigurationOptions): Promise<V1Job>;
-}
-
-/** One RFC 6902 operation used to release an exact suspended Job. */
-interface AgentControllerJobPatchOperation
-{
-	/** Conditional test or one of the two bounded release replacements. */
-	readonly op: "test" | "replace";
-	/** Exact immutable, deadline, or suspend field addressed by the operation. */
-	readonly path: "/metadata/uid" | "/metadata/resourceVersion" | "/spec/activeDeadlineSeconds" | "/spec/suspend";
-	/** Expected field value or bounded release replacement. */
-	readonly value: string | number | boolean;
-}
-
-/** Narrow request that makes JSON Patch semantics explicit at the Kubernetes adapter boundary. */
-interface AgentControllerJobPatchRequest
-{
-	/** Namespace containing the exact assigned Job. */
-	readonly namespace: string;
-	/** Deterministic assigned Job name. */
-	readonly name: string;
-	/** Conditional patch operations in required compare-and-swap order. */
-	readonly body: readonly AgentControllerJobPatchOperation[];
-}
-
-/** Narrow Core API surface used only for exact Pod listing and attempt-key Secret creation. */
-export interface AgentControllerCoreApi
-{
-	/** List Pods using the exact attempt and Kubernetes Job UID selector. */
-	listNamespacedPod(request: { readonly namespace: string; readonly labelSelector: string }, options?: ConfigurationOptions): Promise<V1PodList>;
-	/** Create one immutable, Job-owned attempt-key Secret in the runtime namespace (create-only Role). */
-	createNamespacedSecret(request: { readonly namespace: string; readonly body: V1Secret }, options?: ConfigurationOptions): Promise<V1Secret>;
-}
-
-/** Clients required by the Kubernetes adapter. */
-export interface AgentControllerKubernetesStoreOptions
-{
-	/** Kubernetes Batch client limited by Role to Jobs. */
-	readonly batchApi: AgentControllerBatchApi;
-	/** Kubernetes Core client limited by Role to Pod list. */
-	readonly coreApi: AgentControllerCoreApi;
-	/** Hard timeout independently applied to every Kubernetes request. */
-	readonly requestTimeoutMilliseconds: number;
-	/** Process-lifetime cancellation propagated into every Kubernetes request. */
-	readonly shutdownSignal: AbortSignal;
-}
