+++ libs/backend/agents/runtime/controller/src/index.ts
@@ -1,4 +1,7 @@
-export { __ReconcileNextAgentRuntimeAttempt, __ReconcileNextRuntimeRelease, __RunAgentController, __ValidateAgentControllerRuntimeProfiles } from "./agent-controller.js";
+export { __RunAgentController } from "./agent-controller.js";
+export { __ValidateAgentControllerRuntimeProfiles } from "./agent-controller-profiles.js";
+export { __ReconcileNextAgentRuntimeAttempt } from "./agent-runtime-attempt-assignment.js";
+export { __ReconcileNextRuntimeRelease } from "./agent-runtime-release.js";
 export { __CreateHttpAgentControllerAuthority } from "./http-agent-controller-authority.js";
 export { __CreateKubernetesAgentControllerStore } from "./kubernetes-agent-controller-store.js";
-export type { AgentControllerAuthority, AgentControllerBatchApi, AgentControllerCoreApi, AgentControllerFetch, AgentControllerHttpAuthorityOptions, AgentControllerKubernetesStore, AgentControllerKubernetesStoreOptions, AgentControllerOptions, AgentControllerReconcileResult, AgentControllerRuntimeProfiles, AgentControllerRuntimeReleaseReconcileResult, AgentControllerTokenReader } from "./agent-controller.types.js";
+export type { AgentControllerRuntimeProfiles } from "./agent-controller.types.js";
