+++ libs/backend/agents/runtime/k8s-launcher/src/index.ts
@@ -1,2 +1,5 @@
-export { __AgentRuntimeAttemptResourceName, __BuildSuspendedAgentRuntimeJob, __DeriveAgentRuntimeReleaseDeadlineSeconds } from "./agent-runtime-job.js";
-export type { AgentRuntimeImagePullPolicy, AgentRuntimeJobAssignment, AgentRuntimeJobProfile } from "./agent-runtime-job.types.js";
+export { __BuildSuspendedAgentRuntimeJob } from "./agent-runtime-job.js";
+export { AgentRuntimeIdentityProfiles } from "./agent-runtime-job.types.js";
+export type { AgentRuntimeJobProfile } from "./agent-runtime-job.types.js";
+export { __DeriveAgentRuntimeReleaseDeadlineSeconds } from "./agent-runtime-release-deadline.js";
+export { __AgentRuntimeAttemptResourceName } from "./agent-runtime-resource-name.js";
